refactor: extract resolveHref
